Revert "Move shaded package for Arrow (#362)"
This reverts commit 68236ff8f7e5aa8eaacdea024a3799899282b7fc.
