+++ src/main/java/io/github/zhztheplayer/velox4j/jni/StaticJniApi.java
@@ -33,8 +33,8 @@ import io.github.zhztheplayer.velox4j.query.SerialTaskStats;
 import io.github.zhztheplayer.velox4j.serde.Serde;
 import io.github.zhztheplayer.velox4j.serializable.ISerializable;
 import io.github.zhztheplayer.velox4j.serializable.ISerializableCo;
-import io.github.zhztheplayer.velox4j.shaded.org.apache.arrow.c.ArrowArray;
-import io.github.zhztheplayer.velox4j.shaded.org.apache.arrow.c.ArrowSchema;
+import io.github.zhztheplayer.velox4j.shaded.arrow.org.apache.arrow.c.ArrowArray;
+import io.github.zhztheplayer.velox4j.shaded.arrow.org.apache.arrow.c.ArrowSchema;
 import io.github.zhztheplayer.velox4j.type.RowType;
 import io.github.zhztheplayer.velox4j.type.Type;
 import io.github.zhztheplayer.velox4j.variant.Variant;
