+++ src/main/java/io/github/zhztheplayer/velox4j/shaded/arrow/org/apache/arrow/c/ArrowArray.java
@@ -14,9 +14,9 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-package io.github.zhztheplayer.velox4j.shaded.org.apache.arrow.c;
+package io.github.zhztheplayer.velox4j.shaded.arrow.org.apache.arrow.c;
 
-import static io.github.zhztheplayer.velox4j.shaded.org.apache.arrow.c.NativeUtil.NULL;
+import static io.github.zhztheplayer.velox4j.shaded.arrow.org.apache.arrow.c.NativeUtil.NULL;
 import static org.apache.arrow.util.Preconditions.checkNotNull;
 
 import java.nio.ByteBuffer;
@@ -26,56 +26,61 @@ import org.apache.arrow.memory.ArrowBuf;
 import org.apache.arrow.memory.BufferAllocator;
 import org.apache.arrow.memory.ReferenceManager;
 import org.apache.arrow.memory.util.MemoryUtil;
+import org.apache.arrow.util.VisibleForTesting;
 
-import io.github.zhztheplayer.velox4j.shaded.org.apache.arrow.c.jni.JniWrapper;
+import io.github.zhztheplayer.velox4j.shaded.arrow.org.apache.arrow.c.jni.JniWrapper;
 
 /**
- * C Data Interface ArrowSchema.
+ * C Data Interface ArrowArray.
  *
  * <p>Represents a wrapper for the following C structure:
  *
  * <pre>
- * struct ArrowSchema {
- *     // Array type description
- *     const char* format;
- *     const char* name;
- *     const char* metadata;
- *     int64_t flags;
+ * struct ArrowArray {
+ *     // Array data description
+ *     int64_t length;
+ *     int64_t null_count;
+ *     int64_t offset;
+ *     int64_t n_buffers;
  *     int64_t n_children;
- *     struct ArrowSchema** children;
- *     struct ArrowSchema* dictionary;
+ *     const void** buffers;
+ *     struct ArrowArray** children;
+ *     struct ArrowArray* dictionary;
  *
  *     // Release callback
- *     void (*release)(struct ArrowSchema*);
+ *     void (*release)(struct ArrowArray*);
  *     // Opaque producer-specific data
  *     void* private_data;
  * };
  * </pre>
  */
-public class ArrowSchema implements BaseStruct {
-  private static final int SIZE_OF = 72;
+public class ArrowArray implements BaseStruct {
+  private static final int SIZE_OF = 80;
+  private static final int INDEX_RELEASE_CALLBACK = 64;
 
   private ArrowBuf data;
 
-  /** Snapshot of the ArrowSchema raw data. */
+  /** Snapshot of the ArrowArray raw data. */
   public static class Snapshot {
-    public long format;
-    public long name;
-    public long metadata;
-    public long flags;
+    public long length;
+    public long null_count;
+    public long offset;
+    public long n_buffers;
     public long n_children;
+    public long buffers;
     public long children;
     public long dictionary;
     public long release;
     public long private_data;
 
-    /** Initialize empty ArrowSchema snapshot. */
+    /** Initialize empty ArrowArray snapshot. */
     public Snapshot() {
-      format = NULL;
-      name = NULL;
-      metadata = NULL;
-      flags = NULL;
+      length = NULL;
+      null_count = NULL;
+      offset = NULL;
+      n_buffers = NULL;
       n_children = NULL;
+      buffers = NULL;
       children = NULL;
       dictionary = NULL;
       release = NULL;
@@ -84,45 +89,52 @@ public class ArrowSchema implements BaseStruct {
   }
 
   /**
-   * Create ArrowSchema from an existing memory address.
+   * Create ArrowArray from an existing memory address.
    *
-   * <p>The resulting ArrowSchema does not own the memory.
+   * <p>The resulting ArrowArray does not own the memory.
    *
    * @param memoryAddress Memory address to wrap
-   * @return A new ArrowSchema instance
+   * @return A new ArrowArray instance
    */
-  public static ArrowSchema wrap(long memoryAddress) {
-    return new ArrowSchema(
-        new ArrowBuf(ReferenceManager.NO_OP, null, ArrowSchema.SIZE_OF, memoryAddress));
+  public static ArrowArray wrap(long memoryAddress) {
+    return new ArrowArray(
+        new ArrowBuf(ReferenceManager.NO_OP, null, ArrowArray.SIZE_OF, memoryAddress));
   }
 
   /**
-   * Create ArrowSchema by allocating memory.
+   * Create ArrowArray by allocating memory.
    *
-   * <p>The resulting ArrowSchema owns the memory.
+   * <p>The resulting ArrowArray owns the memory.
    *
    * @param allocator Allocator for memory allocations
-   * @return A new ArrowSchema instance
+   * @return A new ArrowArray instance
    */
-  public static ArrowSchema allocateNew(BufferAllocator allocator) {
-    return new ArrowSchema(allocator.buffer(ArrowSchema.SIZE_OF));
+  public static ArrowArray allocateNew(BufferAllocator allocator) {
+    ArrowArray array = new ArrowArray(allocator.buffer(ArrowArray.SIZE_OF));
+    array.markReleased();
+    return array;
   }
 
-  ArrowSchema(ArrowBuf data) {
-    checkNotNull(data, "ArrowSchema initialized with a null buffer");
+  ArrowArray(ArrowBuf data) {
+    checkNotNull(data, "ArrowArray initialized with a null buffer");
     this.data = data;
+  }
+
+  /** Mark the array as released. */
+  public void markReleased() {
+    directBuffer().putLong(INDEX_RELEASE_CALLBACK, NULL);
   }
 
   @Override
   public long memoryAddress() {
-    checkNotNull(data, "ArrowSchema is already closed");
+    checkNotNull(data, "ArrowArray is already closed");
     return data.memoryAddress();
   }
 
   @Override
   public void release() {
     long address = memoryAddress();
-    JniWrapper.get().releaseSchema(address);
+    JniWrapper.get().releaseArray(address);
   }
 
   @Override
@@ -133,24 +145,30 @@ public class ArrowSchema implements BaseStruct {
     }
   }
 
+  @VisibleForTesting
+  boolean isClosed() {
+    return data == null;
+  }
+
   private ByteBuffer directBuffer() {
-    return MemoryUtil.directBuffer(memoryAddress(), ArrowSchema.SIZE_OF)
+    return MemoryUtil.directBuffer(memoryAddress(), ArrowArray.SIZE_OF)
         .order(ByteOrder.nativeOrder());
   }
 
   /**
-   * Take a snapshot of the ArrowSchema raw values.
+   * Take a snapshot of the ArrowArray raw values.
    *
    * @return snapshot
    */
   public Snapshot snapshot() {
     ByteBuffer data = directBuffer();
     Snapshot snapshot = new Snapshot();
-    snapshot.format = data.getLong();
-    snapshot.name = data.getLong();
-    snapshot.metadata = data.getLong();
-    snapshot.flags = data.getLong();
+    snapshot.length = data.getLong();
+    snapshot.null_count = data.getLong();
+    snapshot.offset = data.getLong();
+    snapshot.n_buffers = data.getLong();
     snapshot.n_children = data.getLong();
+    snapshot.buffers = data.getLong();
     snapshot.children = data.getLong();
     snapshot.dictionary = data.getLong();
     snapshot.release = data.getLong();
@@ -158,14 +176,15 @@ public class ArrowSchema implements BaseStruct {
     return snapshot;
   }
 
-  /** Write values from Snapshot to the underlying ArrowSchema memory buffer. */
+  /** Write values from Snapshot to the underlying ArrowArray memory buffer. */
   public void save(Snapshot snapshot) {
     directBuffer()
-        .putLong(snapshot.format)
-        .putLong(snapshot.name)
-        .putLong(snapshot.metadata)
-        .putLong(snapshot.flags)
+        .putLong(snapshot.length)
+        .putLong(snapshot.null_count)
+        .putLong(snapshot.offset)
+        .putLong(snapshot.n_buffers)
         .putLong(snapshot.n_children)
+        .putLong(snapshot.buffers)
         .putLong(snapshot.children)
         .putLong(snapshot.dictionary)
         .putLong(snapshot.release)
