+++ src/main/java/io/github/zhztheplayer/velox4j/shaded/arrow/org/apache/arrow/c/ArrowArrayStream.java
@@ -14,9 +14,9 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-package io.github.zhztheplayer.velox4j.shaded.org.apache.arrow.c;
+package io.github.zhztheplayer.velox4j.shaded.arrow.org.apache.arrow.c;
 
-import static io.github.zhztheplayer.velox4j.shaded.org.apache.arrow.c.NativeUtil.NULL;
+import static io.github.zhztheplayer.velox4j.shaded.arrow.org.apache.arrow.c.NativeUtil.NULL;
 import static org.apache.arrow.util.Preconditions.checkNotNull;
 
 import java.io.IOException;
@@ -28,8 +28,8 @@ import org.apache.arrow.memory.BufferAllocator;
 import org.apache.arrow.memory.ReferenceManager;
 import org.apache.arrow.memory.util.MemoryUtil;
 
-import io.github.zhztheplayer.velox4j.shaded.org.apache.arrow.c.jni.CDataJniException;
-import io.github.zhztheplayer.velox4j.shaded.org.apache.arrow.c.jni.JniWrapper;
+import io.github.zhztheplayer.velox4j.shaded.arrow.org.apache.arrow.c.jni.CDataJniException;
+import io.github.zhztheplayer.velox4j.shaded.arrow.org.apache.arrow.c.jni.JniWrapper;
 
 /**
  * C Stream Interface ArrowArrayStream.
