+++ src/main/java/io/github/zhztheplayer/velox4j/shaded/arrow/org/apache/arrow/c/jni/JniWrapper.java
@@ -14,7 +14,7 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-package io.github.zhztheplayer.velox4j.shaded.org.apache.arrow.c.jni;
+package io.github.zhztheplayer.velox4j.shaded.arrow.org.apache.arrow.c.jni;
 
 /** JniWrapper for C Data Interface API implementation. */
 public class JniWrapper {
